+++ packages/bonsoir_darwin/darwin/bonsoir_darwin/Sources/bonsoir_darwin/BonsoirService.swift
@@ -13,8 +13,8 @@ class BonsoirService: NSObject {
     /// The response service port.
     var port: Int
     
-    /// The response service name.
-    var host: String?
+    /// The response service host address.
+    var hostAddress: String?
     
     /// The response service mDNS hostname.
     var hostname: String?
@@ -23,11 +23,11 @@ class BonsoirService: NSObject {
     var attributes: [String: String]
     
     /// Creates a new service instance.
-    public init(name: String, type: String, port: Int, host: String?, hostname: String? = nil, attributes: [String: String]) {
+    public init(name: String, type: String, port: Int, hostAddress: String?, hostname: String? = nil, attributes: [String: String]) {
         self.name = name
         self.type = type
         self.port = port
-        self.host = host
+        self.hostAddress = hostAddress
         self.hostname = hostname
         self.attributes = attributes
     }
@@ -38,7 +38,7 @@ class BonsoirService: NSObject {
             "\(prefix)name": name,
             "\(prefix)type": type,
             "\(prefix)port": port,
-            "\(prefix)host": host,
+            "\(prefix)hostAddress": hostAddress,
             "\(prefix)hostname": hostname,
             "\(prefix)attributes": attributes
         ]