+++ packages/bonsoir_windows/windows/bonsoir_service.h
@@ -12,7 +12,7 @@ namespace bonsoir_windows {
     std::string name;
     std::string type;
     int port;
-    std::optional<std::string> host;
+    std::optional<std::string> hostAddress;
     std::optional<std::string> hostname;
     std::map<std::string, std::string> attributes;
 
@@ -20,7 +20,7 @@ namespace bonsoir_windows {
       std::string _name,
       std::string _type,
       int _port,
-      std::optional<std::string> host,
+      std::optional<std::string> hostAddress,
       std::optional<std::string> hostname,
       std::map<std::string, std::string> _attributes
     );